improve setting replacer
The drop list from engine is in reverse order, so e.g. when
trying to set to leaves without leaves in main inventory,
replacer would be set to sapling.
This commit checks all drops first and if finds same drop
as node, it is used. Otherwise the first in list is used.