fix multinode trouble in field mode
when clicking on the 'wrong' part of a multinode-node,
direction calculation is nerfed. This patch avoids a
server crash and emits a sound and message.
User can simply click on another part of the multinode-
node and the procedure should work.

I didn't bother adjusting the pointed_thing to make this work, as I
don't consider the trouble worthwhile for this
kind of node that breaks the voxel world.

The bug was detected by player frogTheSecond and reported
by Huhhila on pandorabox. Thanks guys.